+++ src/main/java/com/wanted/matitnyam/domain/xmlparser/RestaurantsDataParser.java
@@ -1,6 +1,5 @@
 package com.wanted.matitnyam.domain.xmlparser;
 
-import com.wanted.matitnyam.domain.RestaurantsData;
 import jakarta.xml.bind.JAXBContext;
 import jakarta.xml.bind.JAXBException;
 import jakarta.xml.bind.Unmarshaller;
